pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager"
authors = ["Evgeny <info@tinamagic.lv>"]
readme = "README.md"
packages = [{include = "task_manager"}]


[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
django = "^5.1.2"
gunicorn = "^23.0.0"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
django-bootstrap5 = "^24.3"
i18n = "^0.2"
django-filter = "^24.3"
djangorestframework-jsonapi = "^7.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
